refactor(schemas): migrate Phase 12 endpoint schemas — 3 analytics files, 18 classes + 11 enums (#6042)
- analytics_code_generation.py: migrate CodeGenerationRequest, RefactoringRequest,
  ValidationRequest→CodeGenValidationRequest, RollbackRequest→CodeGenRollbackRequest,
  CodeGenerationResponse, RefactoringResponse → schemas_analytics.py;
  also migrate enums RefactoringType, CodeLanguage, GenerationStatus
- analytics_pattern_learning.py: migrate PatternFeedback, PatternDefinition,
  ConfidenceScore, LearningMetrics→PatternLearningMetrics, ActiveLearningQuery,
  PatternUpdate → schemas_analytics.py; also migrate enums FeedbackType,
  PatternCategory, LearningPhase, ConfidenceLevel
- analytics_continuous_learning.py: migrate LearningEvent, Insight→LearningInsight,
  LearningMetrics→ContinuousLearningMetrics, MonitoringStatus→LearningMonitoringStatus,
  RetrainingRequest, LearningConfig → schemas_analytics.py; also migrate enums
  LearningEventType, MonitoringState, InsightType, RetrainingReason

Renames to avoid generic-name collisions:
- ValidationRequest → CodeGenValidationRequest
- RollbackRequest → CodeGenRollbackRequest
- LearningMetrics (in pattern_learning) → PatternLearningMetrics
- LearningMetrics (in continuous_learning) → ContinuousLearningMetrics
- Insight → LearningInsight
- MonitoringStatus → LearningMonitoringStatus
